Добавил защиту от произвольного удаления: при попытке удалить сообщение с предложением через пкм оно будет переотправлено. Взамен требуется удалять через список ответов на предложения, открываемый через контекстное меню.

+++ src/main/java/ru/miau/ilyushabot/functions/private_vcs/VoiceChannelListener.java
@@ -31,7 +31,7 @@ public class VoiceChannelListener extends ListenerAdapter {
 
     private void onJoinVC(GuildVoiceUpdateEvent e) {
         AudioChannelUnion channel = e.getChannelJoined();
-        if (channel.getIdLong() == voiceFabricId.longValue()) {
+        if (channel.getIdLong() == voiceFabricId) {
             Category category = channel.getParentCategory();
             Member member = e.getMember();
             VoiceChannel voiceChannel = category.createVoiceChannel(member.getEffectiveName())
